i686 Debian Stretch Patch for Swift
This gets swift into building 4.2 on x86/i686 Debian Stretch.

The purpose of this is to be an analogue for Raspbian Stretch so 32-bit breaks can be found a bit faster than it would on an armv7 board like the Pi 3B+.

It literally means being able to build in about 1/4th the time when using something like a 2015 iMac with 4 cores available.